Split CLI and self-test surfaces
