+++ Sources/CrawlBarCLI/CLI.swift
@@ -330,122 +330,6 @@ enum CrawlBarCLI {
         return statusService.status(for: installation, timeoutSeconds: 30)
     }
 
-    private static func runConfig(_ options: CLIOptions, registry: CrawlAppRegistry) throws {
-        let store = CrawlBarConfigStore()
-        let nativeConfigStore = CrawlNativeConfigStore()
-        switch options.positionals.first {
-        case "path":
-            print(store.fileURL.path)
-        case "validate":
-            _ = try store.loadOrCreateDefault()
-            print("ok")
-        case "init", nil:
-            let config = try store.loadOrCreateDefault()
-            try CLIOutput.writeJSON(config)
-        case "get":
-            let appID = try options.requiredAppID()
-            let config = try store.loadOrCreateDefault(includeSecrets: options.revealSecrets)
-            let installation = try registry.installation(for: appID)
-            let baseAppConfig = config.appConfig(for: appID) ?? CrawlBarAppConfig(id: appID)
-            let appConfig = installation.map {
-                var copy = baseAppConfig
-                copy.configValues = nativeConfigStore.resolvedConfigValues(appConfig: baseAppConfig, manifest: $0.manifest)
-                return copy
-            } ?? baseAppConfig
-            let values = Self.configValues(
-                appConfig: appConfig,
-                manifest: installation?.manifest,
-                key: options.key,
-                revealSecrets: options.revealSecrets)
-            if options.json {
-                try CLIOutput.writeJSON(values)
-                return
-            }
-            if let key = options.key {
-                guard let value = values.first else {
-                    throw CLIError.usage("unknown config key for \(appID.rawValue): \(key)")
-                }
-                print(value.value ?? "")
-                return
-            }
-            for value in values {
-                print("\(value.id)\t\(value.value ?? "")")
-            }
-        case "set":
-            let appID = try options.requiredAppID()
-            guard let key = options.key?.nilIfBlank else {
-                throw CLIError.usage("config set requires --key <id>")
-            }
-            guard let value = options.value else {
-                throw CLIError.usage("config set requires --value <value>")
-            }
-            var config = try store.loadOrCreateDefault()
-            guard let index = config.apps.firstIndex(where: { $0.id == appID }) else {
-                throw CLIError.usage("unknown app: \(appID.rawValue)")
-            }
-            if value.nilIfBlank == nil {
-                config.apps[index].configValues.removeValue(forKey: key)
-            } else {
-                config.apps[index].configValues[key] = value
-            }
-            let clearMissingSecretIDsByAppID: [CrawlAppID: Set<String>] = value.nilIfBlank == nil ? [appID: [key]] : [:]
-            try store.save(config, clearMissingSecretIDsByAppID: clearMissingSecretIDsByAppID)
-            if let installation = try registry.installation(for: appID),
-               let appConfig = config.appConfig(for: appID)
-            {
-                var nativeAppConfig = appConfig
-                var resolvedValues = nativeConfigStore.resolvedConfigValues(
-                    appConfig: appConfig,
-                    manifest: installation.manifest)
-                if value.nilIfBlank == nil {
-                    resolvedValues.removeValue(forKey: key)
-                } else {
-                    resolvedValues[key] = value
-                }
-                nativeAppConfig.configValues = resolvedValues
-                try nativeConfigStore.write(
-                    appConfig: nativeAppConfig,
-                    manifest: installation.manifest,
-                    clearMissingSecretIDs: clearMissingSecretIDsByAppID[appID] ?? [])
-            }
-            if options.json {
-                try CLIOutput.writeJSON(["app_id": appID.rawValue, "key": key, "updated": "true"])
-                return
-            }
-            print("ok")
-        case let command?:
-            throw CLIError.usage("unknown config command: \(command)")
-        }
-    }
-
-    private static func configValues(
-        appConfig: CrawlBarAppConfig,
-        manifest: CrawlAppManifest?,
-        key: String?,
-        revealSecrets: Bool)
-        -> [CLIConfigValue]
-    {
-        let options = manifest?.configOptions ?? []
-        let knownIDs = Set(options.map(\.id))
-        let extraOptions = appConfig.configValues.keys
-            .filter { !knownIDs.contains($0) }
-            .sorted()
-            .map { CrawlAppManifest.ConfigOption(id: $0, label: $0) }
-        return (options + extraOptions)
-            .filter { key == nil || $0.id == key }
-            .map { option in
-                let rawValue = appConfig.configValues[option.id] ?? option.defaultValue
-                let isSecret = option.kind == .secret
-                return CLIConfigValue(
-                    id: option.id,
-                    label: option.label,
-                    value: isSecret && !revealSecrets && rawValue?.nilIfBlank != nil ? "********" : rawValue,
-                    secret: isSecret,
-                    envVar: option.envVar,
-                    configKey: option.configKey)
-            }
-    }
-
     private static func printHelp() {
         print("""
         crawlbar commands:
@@ -464,109 +348,5 @@ enum CrawlBarCLI {
           config get --app <id> [--key <id>] [--json] [--reveal]
           config set --app <id> --key <id> --value <value> [--json]
         """)
-    }
-}
-
-private struct CLIConfigValue: Encodable {
-    var id: String
-    var label: String
-    var value: String?
-    var secret: Bool
-    var envVar: String?
-    var configKey: String?
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case label
-        case value
-        case secret
-        case envVar = "env_var"
-        case configKey = "config_key"
-    }
-}
-
-private struct CLIApp: Encodable {
-    var id: String
-    var displayName: String
-    var enabled: Bool
-    var available: Bool
-    var availability: CrawlAppManifest.Availability
-    var binaryPath: String?
-    var configPath: String?
-
-    init(_ installation: CrawlAppInstallation) {
-        self.id = installation.id.rawValue
-        self.displayName = installation.manifest.displayName
-        self.enabled = installation.enabled
-        self.available = installation.binaryPath != nil
-        self.availability = installation.manifest.availability
-        self.binaryPath = installation.binaryPath
-        self.configPath = installation.configPathOverride
-    }
-
-    private enum CodingKeys: String, CodingKey {
-        case id
-        case displayName = "display_name"
-        case enabled
-        case available
-        case availability
-        case binaryPath = "binary_path"
-        case configPath = "config_path"
-    }
-}
-
-private struct CLIOptions {
-    var json = false
-    var appID: CrawlAppID?
-    var key: String?
-    var value: String?
-    var revealSecrets = false
-    var diagnostics = false
-    var positionals: [String] = []
-
-    init(_ arguments: ArraySlice<String>) {
-        var iterator = Array(arguments).makeIterator()
-        while let argument = iterator.next() {
-            switch argument {
-            case "--json":
-                self.json = true
-            case "--app":
-                if let value = iterator.next() {
-                    self.appID = CrawlAppID(rawValue: value)
-                }
-            case "--key":
-                self.key = iterator.next()
-            case "--value":
-                self.value = iterator.next()
-            case "--reveal":
-                self.revealSecrets = true
-            case "--diagnostics":
-                self.diagnostics = true
-            case "--":
-                while let value = iterator.next() {
-                    self.positionals.append(value)
-                }
-            default:
-                self.positionals.append(argument)
-            }
-        }
-    }
-
-    func requiredAppID() throws -> CrawlAppID {
-        guard let appID else {
-            throw CLIError.usage("--app <id> is required")
-        }
-        return appID
-    }
-}
-
-private enum CLIError: LocalizedError {
-    case usage(String)
-
-    var errorDescription: String? {
-        switch self {
-        case let .usage(message):
-            message
-        }
     }
 }
